# Instead of appending all of each image values, fix it to extend items…