… at once

Where I follow the example codes in mnist_client.py, I found more efficient way to append image data to request.

As fixing code like below, about 2x faster in the loop. (350ms -> 190ms, In the loop that starts from 91th line on the jupyter notebook)

And the user time of entire built source is decreased 4.5 sec to 3.2 sec.

So I suggest fixing it!