@kirilg 2016-09-10
Thanks for the PR! We've recently changed all of the examples in tensorflow_serving/example (including mnist) to use the standard [model_server](https://github.com/tensorflow/serving/blob/master/tensorflow_serving/model_servers/main.cc) instead of custom servers for each example. As a result, the request type is now [PredictRequest](https://github.com/tensorflow/serving/blob/master/tensorflow_serving/apis/predict.proto), so this change is no longer compatible with the latest client at head.

Please give the new client a try and let us know if you encounter any issues.